ci: update github actions (#456)
- make proto linting mandatory
- check that proto files are up to date
- move docs to the root
- improve the output for doc generation (no error when docs have not been created before)
- remove macOS test environment (very expensive and at the moment unnecessary)

+++ cmd/axelarcli/genDocs.sh
@@ -1,7 +1,10 @@
 #!/bin/sh
 
 # remove old docs
-rm "$1"/*.md
+if find "$1" -name "*.md" 2> /dev/null | grep -q .; then
+  rm "$1"/*.md
+fi
+
 # generate docs
 go run ./ -docs "$1"
 # ensure docs are canonically formatted
